Updated to Complete My Parts
Made a few new tabs and completed my part for this first part. @Kerry, I added a column that may be useful to you. It is in Blue uner "Brian Working Data" It is the sum of services they have. Mike mentioned it in class to me when I was chatting with him and it may be useful to us